fix addFragmentToInfo
