+++ src/utils/addFragmentToInfo.test.js.md
@@ -1,0 +1,243 @@
+# Snapshot report for `dist/utils/addFragmentToInfo.test.js`
+
+The actual snapshot is saved in `addFragmentToInfo.test.js.snap`.
+
+Generated by [AVA](https://ava.li).
+
+## addFragmentToInfo: add field to array payload
+
+> Snapshot 1
+
+    `{␊
+      title␊
+      extraField␊
+    }␊
+    `
+
+> Snapshot 2
+
+    {
+      fieldName: 'books',
+      fragments: {},
+      operation: {
+        kind: 'OperationDefinition',
+        operation: 'query',
+        selectionSet: {
+          kind: 'SelectionSet',
+          selections: [],
+        },
+        variableDefinitions: [],
+      },
+      parentType: 'Query',
+      path: undefined,
+      returnType: '[Book!]!',
+      rootValue: undefined,
+      selectionSet: {
+        kind: 'SelectionSet',
+        selections: {
+          0: {
+            alias: undefined,
+            arguments: {},
+            directives: {},
+            kind: 'Field',
+            name: {
+              kind: 'Name',
+              value: 'title',
+            },
+            selectionSet: undefined,
+          },
+        },
+      },
+      variableValues: {},
+    }
+
+> Snapshot 3
+
+    {
+      fieldName: 'books',
+      fragments: {},
+      operation: {
+        kind: 'OperationDefinition',
+        operation: 'query',
+        selectionSet: {
+          kind: 'SelectionSet',
+          selections: [],
+        },
+        variableDefinitions: [],
+      },
+      parentType: 'Query',
+      path: undefined,
+      returnType: '[Book!]!',
+      rootValue: undefined,
+      selectionSet: {
+        kind: 'SelectionSet',
+        selections: {
+          0: {
+            alias: undefined,
+            arguments: {},
+            directives: {},
+            kind: 'Field',
+            name: {
+              kind: 'Name',
+              value: 'title',
+            },
+            selectionSet: undefined,
+          },
+          1: {
+            alias: undefined,
+            arguments: {},
+            directives: {},
+            kind: 'Field',
+            name: {
+              kind: 'Name',
+              value: 'extraField',
+            },
+            selectionSet: undefined,
+          },
+        },
+      },
+      variableValues: {},
+    }
+
+## makeSubInfo: works when path has been selected
+
+> Snapshot 1
+
+    `{␊
+      content␊
+      wordCount␊
+    }␊
+    `
+
+> Snapshot 2
+
+    {
+      fieldName: 'page',
+      fragments: {},
+      operation: {
+        kind: 'OperationDefinition',
+        operation: 'page',
+        selectionSet: {
+          kind: 'SelectionSet',
+          selections: [],
+        },
+        variableDefinitions: [],
+      },
+      parentType: 'Book',
+      path: {
+        key: 'page',
+        prev: {
+          key: 'book',
+          prev: undefined,
+        },
+      },
+      returnType: 'Page',
+      rootValue: undefined,
+      selectionSet: {
+        kind: 'SelectionSet',
+        selections: {
+          0: {
+            alias: undefined,
+            arguments: {},
+            directives: {},
+            kind: 'Field',
+            name: {
+              kind: 'Name',
+              value: 'content',
+            },
+            selectionSet: undefined,
+          },
+          1: {
+            alias: undefined,
+            arguments: {},
+            directives: {},
+            kind: 'Field',
+            name: {
+              kind: 'Name',
+              value: 'wordCount',
+            },
+            selectionSet: undefined,
+          },
+        },
+      },
+      variableValues: {},
+    }
+
+## makeSubInfo: works when path has been selected and adds fragment
+
+> Snapshot 1
+
+    `{␊
+      content␊
+      wordCount␊
+    }␊
+    `
+
+## makeSubInfo: works with inline fragment
+
+> Snapshot 1
+
+    `{␊
+      page {␊
+        content␊
+      }␊
+    }␊
+    `
+
+> Snapshot 2
+
+    {
+      fieldName: 'page',
+      fragments: {},
+      operation: {
+        kind: 'OperationDefinition',
+        operation: 'page',
+        selectionSet: {
+          kind: 'SelectionSet',
+          selections: [],
+        },
+        variableDefinitions: [],
+      },
+      parentType: 'Book',
+      path: {
+        key: 'page',
+        prev: {
+          key: 'book',
+          prev: undefined,
+        },
+      },
+      returnType: 'Page',
+      rootValue: undefined,
+      selectionSet: {
+        kind: 'SelectionSet',
+        selections: {
+          0: {
+            alias: undefined,
+            arguments: {},
+            directives: {},
+            kind: 'Field',
+            name: {
+              kind: 'Name',
+              value: 'page',
+            },
+            selectionSet: {
+              kind: 'SelectionSet',
+              selections: {
+                0: {
+                  alias: undefined,
+                  arguments: {},
+                  directives: {},
+                  kind: 'Field',
+                  name: {
+                    kind: 'Name',
+                    value: 'content',
+                  },
+                  selectionSet: undefined,
+                },
+              },
+            },
+          },
+        },
+      },
+      variableValues: {},
+    }
